4_Machine_Learning/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
dataclasses==0.8
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
google-auth==2.6.5
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
Markdown==3.3.6
multidict==5.2.0
numpy==1.19.5
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
Pillow==8.1.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytorch-lightning==1.1.4
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
#torch==1.7.1
torchmetrics==0.8.0
#torchvision==0.8.2
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.6.0
